refactor: split hotspot files below 600 LOC with module facades
